+++ webpack/webpack.common.ts
@@ -1,0 +1,31 @@
+import webpack from 'webpack';
+
+const config: webpack.Configuration = {
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js'],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts(x?)$/,
+        exclude: /node_modules/,
+        use: [
+          {
+            loader: 'ts-loader'
+          }
+        ]
+      },
+      {
+        enforce: 'pre',
+        test: /\.js$/,
+        loader: 'source-map-loader'
+      }
+    ]
+  },
+  externals: {
+    'react': 'React',
+    'react-dom': 'ReactDOM'
+  }
+};
+
+export default config;