+++ docs/changelog.md
@@ -6,6 +6,7 @@ Sprint 4 / January 16, 2020
 ### Updates
 - updated readme.
 - created webpack folder to house webpack config files.
+- added webpack.common.ts.
 
 ## Release 0.3.0
 Sprint 3 / January 2, 2020